refactor parser to adhere to a common parse(ctx) interface
Also add much more to context.
Context now contains a `handlers` list that "routes" the data
to the correct layer handlers. This allows for a more flexible
and extensible parser.
Should also be easier to test each layer on it's own.

Have not rewritten a lot of the internal APL logic to follow this yet.
Migth require some more thought. Maybe the internal APL should
have it's own accumulator.