refactor: redid Exmbus.Parser interface.
removed the ParseBehaviour module. We use function captures
as handlers, and there is no way to describe their behaviour other
than that they take a context and return a {:continue, context}
or an {:abort, context} tuple.

Also added a `current_handler` field to the context, and a `handle`
function to the Context.
Now, when `handle(ctx)` is called, it will pop the next handler
and add it to the `current_handler` field.
When an error or warning is pushed to the context, it will
also capture the current handler, so we can know during which handler
the error occurred.