refactor: use prepend_handlers in APL parser to "route" to new handler
APL used to just call down to the FullFrame, FormatFrame,
and CompactFrame parsers directly.
Now, it uses `prepend_handlers` to add the handlers to the context
and then `{:next, ctx}` to "route" to the next handler in the list.

This makes errors from those sub-parsers easier to debug because
the context will attach the associated handler to the error.

By adding the parse calls as handlers, we allow the context
to attach error messages to a more specific handler function.