fix(Tpl): Allow only TPL and APL CI codes in the Tpl parse handler.
Previously, we'd get a FunctionClauseError if we hit an unexpected CI
at the TPL parser. Now we check against our CI lookup table and only
proceed with actual parsing if the CI is a TPL or APL.
We could be even smarter, by adding a CI router in the default handlers,
that would prepend the correct parser module to the context depending
on the layer the CI belongs to.